Move run_pipeline to core module
